+++ golem/browser.py
@@ -5,6 +5,7 @@ import traceback
 from selenium import webdriver
 from selenium.webdriver.remote.webelement import WebElement
 from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
+from selenium.webdriver.chrome.options import Options
 from selenium.common.exceptions import WebDriverException
 
 from golem.core.exceptions import IncorrectSelectorType, ElementNotFound
@@ -237,7 +238,9 @@ def get_browser():
         elif browser_definition['name'] == 'chrome':
             if settings['chromedriver_path']:
                 try:
-                    driver = webdriver.Chrome(executable_path=settings['chromedriver_path'])
+                    chrome_options = Options()
+                    chrome_options.add_argument('--start-maximized')
+                    driver = webdriver.Chrome(executable_path=settings['chromedriver_path'], chrome_options=chrome_options)
                 except:
                     msg = ('Could not start chrome driver using the path \'{}\', '
                            'check the settings file.'.format(settings['chromedriver_path']))
@@ -288,7 +291,6 @@ def get_browser():
                                       desired_capabilities=desired_capabilities,
                                       executable_path=settings['chromedriver_path'])
         elif browser_definition['name'] == 'chrome-remote':
-            print('URL',settings['remote_url'])
             driver = webdriver.Remote(command_executor=settings['remote_url'],
                                       desired_capabilities=DesiredCapabilities.CHROME)
         elif browser_definition['name'] == 'firefox-remote':
@@ -301,7 +303,7 @@ def get_browser():
         
     execution.browser = driver
 
-    # bind _find method to driver instance
+    # bind _find and _find_all methods to driver instance
     driver.find = types.MethodType(_find, driver)
     driver.find_all = types.MethodType(_find_all, driver)