+++ golem/core/test_case.py
@@ -3,25 +3,26 @@ import re
 import sys
 import importlib
 import inspect
+from ast import literal_eval
 
 from golem.core import utils, page_object
 
 
 def _parse_step(step):
-    # ('a')                 -> ['a']
-    # ('a', 'b')            -> ['a', 'b']
-    # ('a, b, c')           -> ['a, b, c']
-    # (('a', 'b', 'c'))     -> ['(\'a\', \'b\', \'c\')']
+    # action('a')                 -> ['a']
+    # action('a', 'b')            -> ['a', 'b']
+    # action('a, b, c')           -> ['a, b, c']
+    # action(('a', 'b', 'c'))     -> ['(\'a\', \'b\', \'c\')']
     method_name = step.split('(', 1)[0].strip()
     if not '.' in method_name:
         method_name = method_name.replace('_', ' ')
-    clean_param_list = []
+    # clean_param_list = []
+    param_list = []
 
     params_re = re.compile('\((?P<args>.+)\)')
     params_search = params_re.search(step)
     if params_search:
         params_string = params_search.group('args')
-        param_list = []
         param_pairs = []
         inside_param = False
         inside_string = False
@@ -38,12 +39,19 @@ def _parse_step(step):
                 elif string_char == '\'':
                     inside_string = False
 
+            if params_string[i] == '\"':
+                if not inside_string:
+                    inside_string = True
+                    string_char = '\"'
+                elif string_char == '\"':
+                    inside_string = False
+
             if params_string[i] == ',' and not inside_param and not inside_string:
                 is_higher_level_comma = True
 
-            if params_string[i] == '(' or params_string[i] == '{':
+            if params_string[i] in ['(', '{', '[']:
                 inside_param = True
-            elif inside_param and (params_string[i] == ')' or params_string[i] == '}'):
+            elif inside_param and params_string[i] in [')', '}', ']']:
                 inside_param = False
 
             if is_higher_level_comma:
@@ -57,19 +65,19 @@ def _parse_step(step):
             param_list.append(params_string[pair[0]:pair[1]])
 
         param_list = [x.strip() for x in param_list]
-        for param in param_list:
-
-            # if 'data[' in param:
-            #     data_re = re.compile("[\'|\"](?P<data>.*)[\'|\"]")
-            #     g = data_re.search(param)
-            #     clean_param_list.append(g.group('data'))
-            if '(' in param and ')' in param:
-                clean_param_list.append(param)
-            else:
-                clean_param_list.append(param.replace('\'', '').replace('"', ''))
+        # for param in param_list:
+        #     # if 'data[' in param:
+        #     #     data_re = re.compile("[\'|\"](?P<data>.*)[\'|\"]")
+        #     #     g = data_re.search(param)
+        #     #     clean_param_list.append(g.group('data'))
+        #     if '(' in param and ')' in param:
+        #         clean_param_list.append(param)
+        #     else:
+        #         # clean_param_list.append(param.replace('\'', '').replace('"', ''))
+        #         clean_param_list.append(param)
     step = {
         'method_name': method_name,
-        'parameters': clean_param_list
+        'parameters': param_list
     }
     return step
 
@@ -102,22 +110,13 @@ def get_test_case_content(project, test_case_name):
 
     test_module = importlib.import_module('projects.{0}.tests.{1}'
                                           .format(project, test_case_name))
-    
     # get description
     description = getattr(test_module, 'description', '')
     
     # get list of pages
     pages = []
     pages = getattr(test_module, 'pages', [])
-    
-    # code_line_list = source_code.split('\n')
-    # for line in code_line_list:
-    #     if 'import' in line:
-    #         import_lines.append(line)
-
-    # for item in inspect.getmembers(test_module):
-    #     print(item[0], isinstance(item[1], types.ModuleType))
-    
+      
     # get setup steps
     setup_steps = []
     setup_function_code = getattr(test_module, 'setup', None)
@@ -145,7 +144,6 @@ def get_test_case_content(project, test_case_name):
         test_contents['content'] = inspect.getsource(test_module)
     except:
         pass
-    print(test_contents['steps'])
     return test_contents
 
 
@@ -187,36 +185,76 @@ def new_test_case(root_path, project, parents, tc_name):
     return errors
 
 
-def format_parameters(step, root_path, project, stored_keys):
-    parameters = step['parameters']
-    action = step['action'].replace(' ', '_')
-    formatted_parameters = []
-    for parameter in parameters:
-        if parameter:
-            this_parameter_string = ''
-            if page_object.is_page_object(parameter, root_path, project):
-                # it is a page object, leave as is
-                this_parameter_string = parameter
-            elif 'random(' in parameter:
-                this_parameter_string = parameter
-            else:
-                is_data_var = 'data.' in parameter or 'data' in parameter
+# class Param_formatter:
 
-                is_in_stored_keys = parameter in stored_keys
-                action_is_store = action == 'store'
-                if (is_data_var or is_in_stored_keys) and not action_is_store:
-                    this_parameter_string = parameter
-                else:
-                    if action in ['wait', 'select_by_index']:
-                        this_parameter_string = parameter
-                    elif parameter[0] == '(' and parameter[-1] == ')':
-                        this_parameter_string = parameter
-                    else:
-                        this_parameter_string = '\'' + parameter + '\''
-            formatted_parameters.append(this_parameter_string)
+#     def __init__(self, root_path, project, stored_keys):
+#         self.root_path = root_path
+#         self.project = project
+#         self.stored_keys = stored_keys
 
-    all_parameters_string = ', '.join(formatted_parameters)
-    return all_parameters_string
+    # def format_param(self, param, action):
+    #     result = ''
+    #     if page_object.is_page_object(param, self.root_path, self.project):
+    #         # it is a page object, leave as is
+    #         result = param
+    #     elif 'random(' in param:
+    #         result = param
+    #     else:
+    #         is_data_var = 'data' in param
+    #         is_in_stored_keys = param in self.stored_keys
+    #         action_is_store = action == 'store'
+    #         if (is_data_var or is_in_stored_keys) and not action_is_store:
+    #             result = param
+    #         else:
+    #             if action in ['wait', 'select_by_index']:
+    #                 result = param
+    #             elif param[0] == '(' and param[-1] == ')':
+    #                 result = param
+    #             else:
+    #                 result = '\'{}\''.format(param)
+    #     return result
+    
+    # def format_param2(self, param, action):
+    #     result = ''
+
+    #     # try:
+    #     #     evalued = literal_eval(param)
+    #     # except:
+    #     #     evalued = None
+
+    #     # if isinstance(evalued, str):
+    #     #     #return '\'{}\''.format(param)
+    #     #     # return '\'' + param + '\''
+    #     #     return param
+    #     # else:
+    #     #     return param
+    #     return param
+
+
+# def format_parameters(param, action, root_path, project, stored_keys):
+#     action = action.replace(' ', '_')
+#     this_parameter_string = ''
+#     if page_object.is_page_object(param, root_path, project):
+#         # it is a page object, leave as is
+#         this_parameter_string = param
+#     elif 'random(' in param:
+#         this_parameter_string = param
+#     else:
+#         is_data_var = 'data.' in param or 'data' in param
+
+#         is_in_stored_keys = param in stored_keys
+#         action_is_store = action == 'store'
+#         if (is_data_var or is_in_stored_keys) and not action_is_store:
+#             this_parameter_string = param
+#         else:
+#             if action in ['wait', 'select_by_index']:
+#                 this_parameter_string = param
+#             elif param[0] == '(' and param[-1] == ')':
+#                 this_parameter_string = param
+#             else:
+#                 this_parameter_string = '\'' + param + '\''
+
+#     return this_parameter_string
 
 
 def format_page_object_string(page_objects):
@@ -241,14 +279,14 @@ def format_description(description):
     return formatted_description
 
 
-def get_stored_keys(steps):
-    stored_keys = []
-    for step in steps:
-        parameters = step['parameters']
-        action = step['action'].replace(' ', '_')
-        if action == 'store':
-            stored_keys.append(parameters[0])
-    return stored_keys
+# def get_stored_keys(steps):
+#     stored_keys = []
+#     for step in steps:
+#         parameters = step['parameters']
+#         action = step['action'].replace(' ', '_')
+#         if action == 'store':
+#             stored_keys.append(parameters[0])
+#     return stored_keys
 
 
 def save_test_case(root_path, project, full_test_case_name, description,
@@ -256,14 +294,18 @@ def save_test_case(root_path, project, full_test_case_name, description,
     tc_name, parents = utils.separate_file_from_parents(full_test_case_name)
     test_case_path = os.path.join(root_path, 'projects', project, 'tests',
                                   os.sep.join(parents), '{}.py'.format(tc_name))
-    setup_stored_keys = get_stored_keys(test_steps['setup'])
-    test_stored_keys = get_stored_keys(test_steps['test'])
-    teardown_stored_keys = get_stored_keys(test_steps['teardown'])
+    # setup_stored_keys = get_stored_keys(test_steps['setup'])
+    # test_stored_keys = get_stored_keys(test_steps['test'])
+    # teardown_stored_keys = get_stored_keys(test_steps['teardown'])
 
+    # stored_keys = setup_stored_keys + test_stored_keys + teardown_stored_keys
 
     formatted_description = format_description(description)
 
+    # param_formatter = Param_formatter(root_path, project, stored_keys)
+    
     with open(test_case_path, 'w', encoding='utf-8') as f:
+        
         # write description
         f.write('\n')
         f.write(formatted_description)
@@ -271,39 +313,49 @@ def save_test_case(root_path, project, full_test_case_name, description,
         # write the list of pages
         f.write('pages = {}\n'.format(format_page_object_string(page_objects)))
         f.write('\n')
+
+
         # write the setup function
         f.write('def setup(data):\n')
         if test_steps['setup']:
             for step in test_steps['setup']:
-                parameters_formatted = format_parameters(step, root_path, project,
-                                                         setup_stored_keys)
+                # params_formatted = []
+                # for param in step['parameters']:
+                #     # params_formatted.append(format_parameter(param, step['action'], root_path, project,
+                #     #                                          setup_stored_keys))
+                #     params_formatted.append(param_formatter.format_param2(param, step['action']))
                 f.write('    {0}({1})\n'
                         .format(step['action'].replace(' ', '_'),
-                                parameters_formatted))
+                                ', '.join(step['parameters'])))
         else:
             f.write('    pass\n')
         f.write('\n')
+        
         # write the test function
         f.write('def test(data):\n')
         if test_steps['test']:
             for step in test_steps['test']:
-                parameters_formatted = format_parameters(step, root_path, project,
-                                                         test_stored_keys)
+                # formatted_params = []
+                # for param in step['parameters']:
+                #     formatted_params.append(format_parameters(param, step['action'],
+                #                                               root_path, project,
+                #                                               test_stored_keys))
                 f.write('    {0}({1})\n'
                         .format(step['action'].replace(' ', '_'),
-                                parameters_formatted))
+                                ', '.join(step['parameters'])))
         else:
             f.write('    pass\n')
         f.write('\n\n')
+        
         # write the teardown function
         f.write('def teardown(data):\n')
         if test_steps['teardown']:
             for step in test_steps['teardown']:
-                parameters_formatted = format_parameters(step, root_path, project,
-                                                         teardown_stored_keys)
+                # parameters_formatted = format_parameters(step, root_path, project,
+                #                                          teardown_stored_keys)
                 f.write('    {0}({1})\n'
                         .format(step['action'].replace(' ', '_'),
-                                parameters_formatted))
+                                ', '.join(step['parameters'])))
         else:
             f.write('    pass\n')