+++ golem/gui/static/js/settings.js
@@ -53,21 +53,13 @@ function saveSettings(){
         type: 'POST',
         success: function(data) {
             if(data.result === 'ok'){
-                toastr.options = {
-                    "positionClass": "toast-top-center",
-                    "timeOut": "2000",
-                    "hideDuration": "100"}
-                toastr.success("Settings saved");
+                utils.toast('success', "Settings saved", 2000);
 
                 if(projectSettingsEditor != null) projectSettingsEditor.markClean();
                 globalSettingsEditor.markClean();
             }
             else{
-                toastr.options = {
-                    "positionClass": "toast-top-center",
-                    "timeOut": "3000",
-                    "hideDuration": "100"}
-                toastr.error(data);
+                utils.toast('error', data, 3000);
             }
         },
         error: function() {